+++ src/voucher_management/__init__.py
@@ -1,3 +1,3 @@
 """Voucher Management application package."""
 
-__version__ = "4.2.0-rc.10"
+__version__ = "4.2.0-rc.12"